+++ src/codemagic/firebase/resources/identifiers/resource_identifier.py
@@ -1,0 +1,12 @@
+from abc import ABC
+from abc import abstractmethod
+from typing import TypeVar
+
+ResourceIdentifierT = TypeVar('ResourceIdentifierT', bound='ResourceIdentifier')
+
+
+class ResourceIdentifier(ABC):
+    @property
+    @abstractmethod
+    def uri(self) -> str:
+        raise NotImplementedError()