+++ src/codemagic/firebase/resources/release.py
@@ -1,19 +1,9 @@
 from dataclasses import dataclass
-from dataclasses import field
 from datetime import datetime
 from datetime import timezone
-from typing import ClassVar
 
-from .resource import DictSerializable
+from .release_notes import ReleaseNotes
 from .resource import Resource
-
-
-@dataclass
-class ReleaseNotes(DictSerializable):
-    """
-    https://firebase.google.com/docs/reference/app-distribution/rest/v1/projects.apps.releases#ReleaseNotes
-    """
-    text: str
 
 
 @dataclass
@@ -29,8 +19,6 @@ class Release(Resource):
     firebaseConsoleUri: str
     testingUri: str
     binaryDownloadUri: str
-
-    label: ClassVar[str] = field(default='releases', init=False, repr=False)
 
     def __post_init__(self):
         if isinstance(self.createTime, str):