+++ src/codemagic/firebase/resources/resource.py
@@ -1,23 +1,31 @@
+from __future__ import annotations
+
 from abc import ABC
-from dataclasses import dataclass
 from datetime import datetime
 from typing import ClassVar
 from typing import Dict
+from typing import Optional
+from typing import TypeVar
+
+ResourceT = TypeVar('ResourceT', bound='Resource')
 
 
-class DictSerializable:
+class Resource(ABC):
+    __google_api_label__: ClassVar[Optional[str]] = None
+
     @classmethod
-    def _serialize(cls, obj):
-        if isinstance(obj, DictSerializable):
-            return obj.dict()
-        if isinstance(obj, datetime):
-            return obj.isoformat()
-        return obj
+    def get_label(cls) -> str:
+        if cls.__google_api_label__:
+            return cls.__google_api_label__
+        return f'{cls.__name__.lower()}s'
 
     def dict(self) -> Dict:
         return {k: self._serialize(v) for k, v in self.__dict__.items()}
 
-
-@dataclass
-class Resource(DictSerializable, ABC):
-    label: ClassVar[str]
+    @classmethod
+    def _serialize(cls, obj):
+        if isinstance(obj, Resource):
+            return obj.dict()
+        if isinstance(obj, datetime):
+            return obj.isoformat()
+        return obj