+++ src/codemagic/tools/firebase/action_groups/releases_action_group.py
@@ -4,10 +4,8 @@ from typing import List
 
 from codemagic import cli
 from codemagic.firebase.api_error import FirebaseApiClientError
-from codemagic.firebase.resource_managers.release_manager import AppId
-from codemagic.firebase.resource_managers.release_manager import ProjectId
-from codemagic.firebase.resource_managers.release_manager import ReleaseParentIdentifier
 from codemagic.firebase.resources import Release
+from codemagic.firebase.resources.identifiers import AppIdentifier
 
 from ..arguments import FirebaseArgument
 from ..arguments import ReleasesArgument
@@ -26,8 +24,8 @@ class ReleasesActionGroup(FirebaseAction, metaclass=ABCMeta):
     )
     def list_releases(
         self,
-        project_id: ProjectId,
-        app_id: AppId,
+        project_id: str,
+        app_id: str,
         json_output: bool = False,
         should_print: bool = True,
     ) -> List[Release]:
@@ -35,9 +33,9 @@ class ReleasesActionGroup(FirebaseAction, metaclass=ABCMeta):
         List releases for the specified project and application from Firebase API
         """
 
-        parent_identifier = ReleaseParentIdentifier(project_id, app_id)
+        app_identifier = AppIdentifier(project_id, app_id)
         try:
-            releases = self.api_client.releases.list(parent_identifier)
+            releases = self.api_client.releases.list(app_identifier)
         except FirebaseApiClientError as e:
             raise FirebaseError(str(e))