+++ src/codemagic/tools/firebase/firebase_action.py
@@ -3,8 +3,6 @@ from abc import ABCMeta
 from abc import abstractmethod
 
 from codemagic.firebase.api_client import FirebaseApiClient
-from codemagic.firebase.resource_managers.release_manager import AppId
-from codemagic.firebase.resource_managers.release_manager import ProjectId
 
 
 class FirebaseAction(metaclass=ABCMeta):
@@ -22,8 +20,8 @@ class FirebaseAction(metaclass=ABCMeta):
     @abstractmethod
     def get_latest_build_version(
         self,
-        project_id: ProjectId,
-        app_id: AppId,
+        project_id: str,
+        app_id: str,
     ) -> int:
         from .actions import GetLatestBuildVersionAction
         _ = GetLatestBuildVersionAction.get_latest_build_version  # Implementation